Menu changes, Save Load game changes, Ammo, Chest, Powerups, UI for powerups

+++ src/main/java/uet/project/arkanoid/objects/MovableObject.java
@@ -5,7 +5,7 @@ import uet.project.arkanoid.base.Vector2D;
 public abstract class MovableObject extends GameObject {
     private Vector2D direction;
 
-    public MovableObject(int x, int y, double width, double height) {
+    public MovableObject(double x, double y, double width, double height) {
         super(x, y, width, height);
         this.direction = new Vector2D(0, 0);
     }
